Move shifting to separate modules
