+++ src/biguint/algorithms.rs
@@ -1,13 +1,8 @@
-use crate::std_alloc::{Cow, Vec};
 use core::cmp::Ordering::{self, Equal};
-use core::iter::repeat;
 use core::mem;
-use num_traits::{PrimInt, Zero};
+use num_traits::PrimInt;
 
-use crate::biguint::biguint_from_vec;
-use crate::biguint::BigUint;
-
-use crate::big_digit::{self, BigDigit};
+use crate::big_digit::BigDigit;
 
 /// Find last set bit
 /// fls(0) == 0, fls(u32::MAX) == 32
@@ -17,89 +12,6 @@ pub(crate) fn fls<T: PrimInt>(v: T) -> u8 {
 
 pub(crate) fn ilog2<T: PrimInt>(v: T) -> u8 {
     fls(v) - 1
-}
-
-#[inline]
-pub(crate) fn biguint_shl<T: PrimInt>(n: Cow<'_, BigUint>, shift: T) -> BigUint {
-    if shift < T::zero() {
-        panic!("attempt to shift left with negative");
-    }
-    if n.is_zero() {
-        return n.into_owned();
-    }
-    let bits = T::from(big_digit::BITS).unwrap();
-    let digits = (shift / bits).to_usize().expect("capacity overflow");
-    let shift = (shift % bits).to_u8().unwrap();
-    biguint_shl2(n, digits, shift)
-}
-
-fn biguint_shl2(n: Cow<'_, BigUint>, digits: usize, shift: u8) -> BigUint {
-    let mut data = match digits {
-        0 => n.into_owned().data,
-        _ => {
-            let len = digits.saturating_add(n.data.len() + 1);
-            let mut data = Vec::with_capacity(len);
-            data.extend(repeat(0).take(digits));
-            data.extend(n.data.iter());
-            data
-        }
-    };
-
-    if shift > 0 {
-        let mut carry = 0;
-        let carry_shift = big_digit::BITS as u8 - shift;
-        for elem in data[digits..].iter_mut() {
-            let new_carry = *elem >> carry_shift;
-            *elem = (*elem << shift) | carry;
-            carry = new_carry;
-        }
-        if carry != 0 {
-            data.push(carry);
-        }
-    }
-
-    biguint_from_vec(data)
-}
-
-#[inline]
-pub(crate) fn biguint_shr<T: PrimInt>(n: Cow<'_, BigUint>, shift: T) -> BigUint {
-    if shift < T::zero() {
-        panic!("attempt to shift right with negative");
-    }
-    if n.is_zero() {
-        return n.into_owned();
-    }
-    let bits = T::from(big_digit::BITS).unwrap();
-    let digits = (shift / bits).to_usize().unwrap_or(core::usize::MAX);
-    let shift = (shift % bits).to_u8().unwrap();
-    biguint_shr2(n, digits, shift)
-}
-
-fn biguint_shr2(n: Cow<'_, BigUint>, digits: usize, shift: u8) -> BigUint {
-    if digits >= n.data.len() {
-        let mut n = n.into_owned();
-        n.set_zero();
-        return n;
-    }
-    let mut data = match n {
-        Cow::Borrowed(n) => n.data[digits..].to_vec(),
-        Cow::Owned(mut n) => {
-            n.data.drain(..digits);
-            n.data
-        }
-    };
-
-    if shift > 0 {
-        let mut borrow = 0;
-        let borrow_shift = big_digit::BITS as u8 - shift;
-        for elem in data.iter_mut().rev() {
-            let new_borrow = *elem << borrow_shift;
-            *elem = (*elem >> shift) | borrow;
-            borrow = new_borrow;
-        }
-    }
-
-    biguint_from_vec(data)
 }
 
 pub(crate) fn cmp_slice(a: &[BigDigit], b: &[BigDigit]) -> Ordering {
